added cast in order to avoid C2036 error under msvc compilation

+++ src/client.c
@@ -12,7 +12,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <errno.h>
-
+#include <inttypes.h>
 #include <rudp/packet.h>
 #include <rudp/client.h>
 #include <rudp/peer.h>
@@ -183,7 +183,7 @@ rudp_error_t rudp_client_send(
 
         pcs[i] = rudp_packet_chain_alloc(client->rudp, bytes_to_write+header_size);
 
-        memcpy(&pcs[i]->packet->data.data[0], data+bytes_written, bytes_to_write);
+        memcpy(&pcs[i]->packet->data.data[0], ((uint8_t)data) + bytes_written, bytes_to_write);
 
         bytes_written += bytes_to_write;
